Improvements of surface smoothing, feature edge capturing + various fixes
git-svn-id: https://pl5.projectlocker.com/igui/meshGeneration/svn@62 fdcce57e-7e00-11e2-b579-49867b4cea03

+++ meshLibrary/utilities/surfaceTools/meshSurfaceMapper/meshSurfaceMapper.C
@@ -55,7 +55,8 @@ void meshSurfaceMapper::createTriSurfacePartitioner() const
     
 void meshSurfaceMapper::clearOut()
 {
-    deleteDemandDrivenData(surfaceEnginePartitionerPtr_);
+    if( deletePartitioner_ )
+        deleteDemandDrivenData(surfaceEnginePartitionerPtr_);
     deleteDemandDrivenData(surfPartitionerPtr_);
 }
 
@@ -70,6 +71,27 @@ meshSurfaceMapper::meshSurfaceMapper
     surfaceEngine_(mse),
     meshOctree_(octree),
     surfaceEnginePartitionerPtr_(NULL),
+    deletePartitioner_(true),
+    surfPartitionerPtr_(NULL)
+{
+    if( Pstream::parRun() )
+    {
+        //- allocate bpAtProcs and other addressing
+        //- this is done here to prevent possible deadlocks
+        surfaceEngine_.bpAtProcs();
+    }
+}
+
+meshSurfaceMapper::meshSurfaceMapper
+(
+    const meshSurfacePartitioner& mPart,
+    const meshOctree& octree
+)
+:
+    surfaceEngine_(mPart.surfaceEngine()),
+    meshOctree_(octree),
+    surfaceEnginePartitionerPtr_(&mPart),
+    deletePartitioner_(false),
     surfPartitionerPtr_(NULL)
 {
     if( Pstream::parRun() )
